python-examples/requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
requests>=2.31.0
